==> DataOrganization/2_add_conditionANDvideo_labels.ipynb <==
import pickle
import time
import os
import datetime
import numpy as np
import pandas as pd

cur_dir = os.getcwd()
main_dir = os.path.dirname(os.path.dirname(cur_dir))

data_dir = os.path.join(main_dir, "ConsolidatedData", "Dynamic1_Jan2023")
with open(os.path.join(data_dir, "nopeaks_consolidated_all.pkl"),'rb') as f:
  df = pickle.load(f)

#For training tests
#Add "Label" column to indicate flight condition where flag==0b10011 (capturing data)
if "Label" not in df.columns:
  df.insert(1, "Label", None)
  airspeeds = [7, 8.3, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20]
  aoas = [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16]

  conds = list()
  for airspeed in airspeeds:
    for aoa in aoas:
      conds.append (f"{airspeed}m/s_{aoa}deg")

  cond_ix = -1
  prev_flag = 0
  for row_ix in range(df.shape[0]):
    if row_ix % 100000 == 0:
      print (row_ix)
    row = df.iloc[row_ix]
    if row["Flag"] == "0b10011":
      if prev_flag == 0:
        cond_ix += 1
        prev_flag = 1
      df.loc[row_ix, "Label"] = conds[cond_ix]
    else:
      prev_flag = 0

#For dynamic tests
#Add "Label" column to indicate flight condition where flag==0b10011 (capturing data)
if "Label" not in df.columns:
  df.insert(1, "Label_red", None)
  df.insert(1, "Label", None)
  
  airspeeds = np.asarray([7, 8.3, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20])
  aoas = np.asarray([0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16])

  for row_ix in range(df.shape[0]):
    if row_ix % 100000 == 0:
      print (row_ix)
    row = df.iloc[row_ix]
    
    airspeed = row["MPH"]*0.44704
    i = (np.abs(airspeeds - airspeed)).argmin()
    airspeed_red = airspeeds[i]

    aoa = row["Inclination (deg)"]
    i = (np.abs(aoas - aoa)).argmin()
    aoa_red = aoas[i]

    df.loc[row_ix, "Label_red"] = f"{airspeed_red}m/s_{aoa_red}deg"
    df.loc[row_ix, "Label"] = f"{airspeed:.2f}m/s_{aoa:.2f}deg"

df.head()

df.tail()

#For training
#Add "Video Time (sec)" column to match the seconds in the run video.
video_start_rows_in_df = [1480]
video_start_seconds_in_video = [5]
video_end_rows_in_df = [1417480]
video_end_seconds_in_video = [10972.0]

df.insert(2, "Video Time (sec)", None)

for i in range(len(video_start_rows_in_df)):
  num_rows = video_end_rows_in_df[i] - video_start_rows_in_df[i]
  time_delta = video_end_seconds_in_video[i] - video_start_seconds_in_video[i]
  video_time_inc = time_delta/num_rows

  time_increment_secs = 0
  video_time = video_start_seconds_in_video[i]
  for j in range(num_rows):
    df.loc[video_start_rows_in_df[i]+j, "Video Time (sec)"] = video_time + time_increment_secs
    time_increment_secs += video_time_inc

#For dynamic
#Remove the first "n" rows at the beginning of dataframe that's not in the 1st processed video
top_n_lines_to_remove = 167507 #Found directly by matching the timestamp at the beginning of first processed video (Run 15) & the EDS data.
df = df.iloc[top_n_lines_to_remove:]

#Record the offset time because we will also remove these lines from SG/RTD and PZT data.
time_obj = df["Date/Time"].apply(lambda x: datetime.datetime.strptime(x, "%Y-%m-%d_%H-%M-%S-%f"))
start_time_dynamic = time_obj.iloc[0]
print (start_time_dynamic)

offsets_savedir = os.path.join(cur_dir, 'offset_pickles')
with open(os.path.join(offsets_savedir,'starttime_dynamic1.pkl'), 'wb') as f:
  pickle.dump(start_time_dynamic, f)

#For dynamic
#Add "Video Time (sec)" column to match the seconds in the processed run videos.
df.insert(3, "Video Time (sec)", None)

init_time = time_obj.iloc[1]
run_num = df.iloc[1]["Run Number"]

for row_ix in range(1,df.shape[0]):
  if row_ix % 100000 == 0:
      print (row_ix)
  row_run_num = df.iloc[row_ix]["Run Number"]
  if run_num == row_run_num:
    time_passed = time_obj.iloc[row_ix] - init_time
    df.iloc[row_ix,df.columns.get_loc("Video Time (sec)")] = time_passed.total_seconds()
  else:
    init_time = time_obj.iloc[row_ix]
    run_num = row_run_num


#Save videotimed and labeled data as .csv and pickle files
output_csv = os.path.join(data_dir, "nopeaks_videotimed_labeled_consolidated_all.csv")
df.to_csv(output_csv, index=False)

df.to_pickle(os.path.join(data_dir,'nopeaks_videotimed_labeled_consolidated_all.pkl'))
